Export functions from each route
Previously, the module.exports of each route was a function that
accepted the app object as a parameter and took care of registering the
endpoint. Now, that function is exported on routeName.endpoint. This
allows other functions to be exported as well, such as the getters and
processors that were previously only available to the route itself.
This will allow routes to access data from other routes with maximal
reuse of code.

This required a change in the way modules are loaded in index.js.
Routes are now specified by name in an array, then dynamically imported
by a for loop, allowing for simpler addition of new routes.