Change from custom to default response formatter
The custom response formatter was created based on the incorrect idea
that Restify did not ship with a default JSON formatter. Apparently, all
that is needed to use the stock JSON formatter is to set the
Content-Type of the response before sending it. This change has the dual
benefit of simplifying the response handler chain and automatically
handling Restify errors.