Improve database interaction
Create a more robust and expressive wrapper for the CouchDB database
that is more in line with industry standards.

Add configuration for database names

Since CouchDB has the concept of multiple databases inside an instance,
there should be support for the app writing to more than one
database. With this change, it would be confusing to have a key `db`
inside an object called `db`, so the whole database configuration hash
is now referenced by the key `couchDB`.

Add generic database names to vars template

Since the app depends on them, it makes sense to supply sensible default
names for the databases.

Refactor database helper into a class

The class approach allows for instantiation of the database in the
functions that use it, instead of centrally. This provides the
flexibility to reference different databases inside the same CouchDB
instance instead of being forced to use one database for everything.

Change "user" property to "username"

Capitalize name of DB class file

Change DB class name to Database

Add models

The root classes are Database and Model. A Model represents a document
in a database. Model classes instantiate a Database with a given name.
User and AppData extend Model, but already know the database name to
instantiate.

Fix model db instantiation and add db error

Both models were failing to pass the configuration object to the
Database constructor, causing it to fail silently to instantiate a
database instance. Now, both models pass the correct parameters and the
Database constructor throws an error if the requested database does not
exist.

Convert routes to use Database class

Instead of passing a function as the getter property, the endpoint
class now receives a Model, which it can call methods on.

Convert cache to use Database class

Add error handling

Fix meal points per day not handling model

This endpoint is different than the rest because it handles getting data
from two separate endpoints, so it overrides the logic that the other
endpoints use, located in the `Endpoint` class. Since days left in
semester now uses a `Model` for data getting, this route needed to be
updated to support the `Model` getter method.