Move route creation logic into dedicated file
Instead of loading routes based on their file being in the routes
directory, load them based on a list defined in a file. This change
moves the logic for loading routes out of `index.js` and into
`/routes/routes.js`, which is a cleaner and more flexible approach. The
tests had to be updated to reflect this change, because they previously
used the same dynamic approach for defining enabled routes, and now need
to get the list of enabled routes from `routes.js`.