Convert all routes to classes
Each route extends the base `Endpoint` class and overrides the getter and
processor functions to provide custom functionality. The constructor
accepts an `app` object that is passed, along with some configuration
settings, up to the base `Endpoint` class, which adds the endpoint to
the Restify app. When the app object is *not* passed, the endpoint class
is still instantiated with all of its properties but is, obviously, not
added to the nonexistent Restify app. This is useful in tests.

The `Endpoint` class was moved from helpers to routes because it makes
more sense to be located on the same folder level as the routes. The
`routes.js` file takes care of defining and loading all of the enabled
routes. Its definition of all enabled routes is useful in tests.