Add request logging and fix caching
Arrays were not being cached because they were being caught as objects,
which they technically are, and assigned to undefined values. Now,
arrays will be cached just like other data.

Each request will be logged in the requesting user's data. If a username
is provided with a GET request, it will be logged.